+++ gcl/o/usig.c
@@ -226,7 +226,7 @@ DEFUN_NEW("FPE_CODE",object,fSfpe_code,SI,2,2,NONE,II,OO,OO,OO,(fixnum x87sw,fix
 DEFUN_NEW("FNSTSW",object,fSfnstsw,SI,0,0,NONE,II,OO,OO,OO,(void),"") {
   volatile unsigned short t;
   ASM ("fnstsw %0" :: "m" (t));
-  RETURN1((object)t);
+  RETURN1((object)(long)t);
 }
 DEFUN_NEW("STMXCSR",object,fSstmxcsr,SI,0,0,NONE,II,OO,OO,OO,(void),"") {
   volatile unsigned int t;
